optimized wall and post
