+++ src/main/kotlin/ru/netology/post/Comments.kt
@@ -1,0 +1,4 @@
+package ru.netology.post
+
+class Comments {
+}
