+++ src/main/kotlin/ru/netology/post/Copyright.kt
@@ -1,0 +1,4 @@
+package ru.netology.post
+
+class Copyright {
+}
