+++ src/main/kotlin/ru/netology/post/Likes.kt
@@ -1,0 +1,4 @@
+package ru.netology.post
+
+class Likes {
+}
