+++ src/main/kotlin/ru/netology/post/Reports.kt
@@ -1,0 +1,4 @@
+package ru.netology.post
+
+class Reports {
+}
