+++ src/main/kotlin/ru/netology/post/Views.kt
@@ -1,0 +1,4 @@
+package ru.netology.post
+
+class Views {
+}
